+++ lib/attributes_dsl/attributes.rb
@@ -10,21 +10,14 @@ module AttributesDSL
   # @author Andrew Kozin <Andrew.Kozin@gmail.com>
   #
   class Attributes
-
-    include Immutability
-
     # @!attribute [r] attributes
     #
     # Uses the set of attributes to ensure their uniqueness (by name)
     #
     # @return [Set] the set of registered attributes
     #
-    attr_reader :attributes
-
-    # Initializes an immutable collection with an initial set of attributes
-    #
-    def initialize
-      @attributes = {}
+    def attributes
+      @attributes ||= {}
     end
 
     # Initializes the attribute from given arguments
@@ -34,40 +27,49 @@ module AttributesDSL
     #
     # @return [AttributesDSL::Attributes]
     #
-    def register(name, options = {}, &coercer)
-      attribute = Attribute.new(name, options, &coercer)
-      update { @attributes = attributes.merge(name => attribute) }
+    def add(name, options = {}, &coercer)
+      name  = name.to_sym
+      value = Attribute.new(name, options, &coercer)
+      clone_with do
+        @attributes  = attributes.merge(name => value)
+        @transformer = nil
+      end
     end
 
-    # Extracts instance attributes from the input hash
+    # Returns the proc that converts a hash of attributes using current setting
     #
-    # Assigns default values and uses coercions when applicable.
+    # @return [Proc]
     #
-    # @param [Hash] input
+    def transformer
+      @transformer ||= transprocs.flatten.compact.reduce(:>>)
+    end
+
+    # Checks whether an attribute reader should be defined
     #
-    # @return [Hash]
+    # @param [#to_sym] name
     #
-    def extract(input)
-      validate(input).inject({}) do |a, e|
-        key = e.name
-        value = input.fetch(key) { e.default }
-        a.merge(key => e.value(value))
-      end
+    # @return [Boolean]
+    #
+    def reader?(name)
+      attributes[name.to_sym].reader
     end
 
     private
 
-    def validate(input)
-      undefined = required - input.keys
-      return attributes.values if undefined.empty?
-
-      fail ArgumentError.new "Undefined attributes: #{undefined.join(", ")}"
+    def transprocs
+      [
+        Transprocs[:filter, keys],
+        attributes.values.map(&:transformer),
+        Transprocs[:update, keys]
+      ]
     end
 
-    def required
-      attributes.values.select(&:required).map(&:name)
+    def keys
+      attributes.keys
     end
 
-  end # class Attributes
-
-end # module AttributesDSL
+    def clone_with(&block)
+      dup.tap { |instance| instance.instance_eval(&block) }
+    end
+  end
+end