+++ spec/integration/unknown_attribute_spec.rb
@@ -1,0 +1,10 @@
+describe "unknown attribute" do
+  let(:klass) { Class.new { extend AttributesDSL } }
+
+  subject { klass.new foo: :BAR }
+
+  it "is ignored" do
+    expect(subject).not_to respond_to :foo
+    expect(subject.attributes).to eql({})
+  end
+end