+++ CHANGELOG.md
@@ -1,14 +1,25 @@
-## version 0.0.3 to be released
+## version 0.1.0 2015-12-08
+
+Stabilized DSL
 
 ### Added
 
 * Option `:reader` that allows skipping attribute reader (nepalez)
+* Option `:only` whitelists allowed values (nepalez)
+* Option `:except` blacklists allowed values (nepalez)
+* Option `:coercer` is an alternative to block syntax (nepalez)
+* An instance can be initialized by hash with both symbolic and string keys (nepalez)
+
+### Deleted
+
+* Support for rubies < 2.1
 
 ### Internal
 
-* Switched to 'immutability' API (nepalez)
+* Removed freezing of values (nepalez)
+* Slightly (20-25%) improved efficiency due to usage of transpfoc (nepalez)
 
-[Compare v0.0.2...HEAD](https://github.com/nepalez/attributes_dsl/compare/v0.0.2...HEAD)
+[Compare v0.0.2...v0.1.0](https://github.com/nepalez/attributes_dsl/compare/v0.0.2...v0.1.0)
 
 ## version 0.0.2 2015-09-11